requirements.txt
fastapi
uvicorn
websockets
pydantic
google-generativeai
openai
